move to depot & fill the whole tile
