+++ src/assets/sass/light-bootstrap-dashboard.css
@@ -1,8 +1,8 @@
 /*!
 
  =========================================================
- * Light Bootstrap Dashboard React - v1.1.0
- * Based on Light Bootstrap Dashboard - v1.3.0
+ * Light Bootstrap Depot React - v1.1.0
+ * Based on Light Bootstrap Depot - v1.3.0
  =========================================================
 
  * Product Page: http://www.creative-tim.com/product/light-bootstrap-dashboard-react
