--- src/views/Dashboard/Cards/index.ts
@@ -1,4 +1,0 @@
-import BalanceTile from './BalanceTile';
-import StockBalanceTile from './StockBalanceTile';
-
-export { BalanceTile, StockBalanceTile };
