--- src/views/Dashboard/index.ts
@@ -1,3 +1,0 @@
-import Dashboard from './Dashboard';
-
-export default Dashboard;
