+++ src/views/Depot/Cards/index.ts
@@ -1,0 +1,5 @@
+import BalanceCard from './BalanceCard';
+import StockBalanceCard from './StockBalanceCard';
+import StockShareCard from './StockShareCard';
+
+export { BalanceCard, StockBalanceCard, StockShareCard };
